Feat: #BBB-156 쿠폰 시스템 적용 & 유저의 포인트 및 신뢰도 증감 로직 구현

+++ domain/src/main/java/com/bombombom/devs/study/model/Study.java
@@ -103,8 +103,8 @@ public abstract class Study extends BaseEntity {
     public abstract StudyType getStudyType();
 
 
-    public Integer calculateDeposit() {
-        return penalty * weeks;
+    public Long calculateDeposit() {
+        return (long) penalty * weeks;
     }
 
     public void admit(User user) {
